+++ Source/HoudiniEngine/Private/UnrealLandscapeSplineTranslator.cpp
@@ -220,7 +220,7 @@ FHoudiniUnResampledPoint::GetSelectedPosition() const
 	case EHoudiniUnrealLandscapeSplineCurve::Right:
 		return Right;
 	default:
-		HOUDINI_LOG_WARNING(TEXT("Invalid value for SplineSelection: %d, returning Center point."), SplineSelection);
+		HOUDINI_LOG_WARNING(TEXT("Invalid value for SplineSelection: %d, returning Center point."), static_cast<int32>(SplineSelection));
 		break;
 	}
