+++ Source/HoudiniEngineEditor/Private/HoudiniPublicAPIAssetWrapper.cpp
@@ -943,7 +943,7 @@ UHoudiniPublicAPIAssetWrapper::SetFloatParameterValue_Implementation(FName InPar
 		if (!IsValid(FloatParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -962,7 +962,7 @@ UHoudiniPublicAPIAssetWrapper::SetFloatParameterValue_Implementation(FName InPar
 		if (!IsValid(ColorParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1010,7 +1010,7 @@ UHoudiniPublicAPIAssetWrapper::GetFloatParameterValue_Implementation(FName InPar
 		if (!IsValid(FloatParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1029,7 +1029,7 @@ UHoudiniPublicAPIAssetWrapper::GetFloatParameterValue_Implementation(FName InPar
 		if (!IsValid(ColorParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1064,7 +1064,7 @@ UHoudiniPublicAPIAssetWrapper::SetColorParameterValue_Implementation(FName InPar
 		if (!IsValid(ColorParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1098,7 +1098,7 @@ UHoudiniPublicAPIAssetWrapper::GetColorParameterValue_Implementation(FName InPar
 		if (!IsValid(ColorParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1125,7 +1125,7 @@ UHoudiniPublicAPIAssetWrapper::SetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(IntParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1144,7 +1144,7 @@ UHoudiniPublicAPIAssetWrapper::SetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(ChoiceParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1156,7 +1156,7 @@ UHoudiniPublicAPIAssetWrapper::SetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(MultiParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1168,7 +1168,7 @@ UHoudiniPublicAPIAssetWrapper::SetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(ToggleParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1180,7 +1180,7 @@ UHoudiniPublicAPIAssetWrapper::SetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(FolderParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1239,7 +1239,7 @@ UHoudiniPublicAPIAssetWrapper::GetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(ChoiceParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1252,7 +1252,7 @@ UHoudiniPublicAPIAssetWrapper::GetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(MultiParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1265,7 +1265,7 @@ UHoudiniPublicAPIAssetWrapper::GetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(ToggleParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1278,7 +1278,7 @@ UHoudiniPublicAPIAssetWrapper::GetIntParameterValue_Implementation(FName InParam
 		if (!IsValid(FolderParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1305,7 +1305,7 @@ UHoudiniPublicAPIAssetWrapper::SetBoolParameterValue_Implementation(FName InPara
 		if (!IsValid(ToggleParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1317,7 +1317,7 @@ UHoudiniPublicAPIAssetWrapper::SetBoolParameterValue_Implementation(FName InPara
 		if (!IsValid(FolderParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1355,7 +1355,7 @@ UHoudiniPublicAPIAssetWrapper::GetBoolParameterValue_Implementation(FName InPara
 		if (!IsValid(ToggleParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1368,7 +1368,7 @@ UHoudiniPublicAPIAssetWrapper::GetBoolParameterValue_Implementation(FName InPara
 		if (!IsValid(FolderParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1395,7 +1395,7 @@ UHoudiniPublicAPIAssetWrapper::SetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(StringParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1439,7 +1439,7 @@ UHoudiniPublicAPIAssetWrapper::SetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(ChoiceParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1452,7 +1452,7 @@ UHoudiniPublicAPIAssetWrapper::SetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(FileParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1493,7 +1493,7 @@ UHoudiniPublicAPIAssetWrapper::GetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(StringParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1530,7 +1530,7 @@ UHoudiniPublicAPIAssetWrapper::GetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(ChoiceParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1544,7 +1544,7 @@ UHoudiniPublicAPIAssetWrapper::GetStringParameterValue_Implementation(FName InPa
 		if (!IsValid(FileParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1578,7 +1578,7 @@ UHoudiniPublicAPIAssetWrapper::SetAssetRefParameterValue_Implementation(FName In
 		if (!IsValid(StringParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1625,7 +1625,7 @@ UHoudiniPublicAPIAssetWrapper::GetAssetRefParameterValue_Implementation(FName In
 		if (!IsValid(StringParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -1673,7 +1673,7 @@ UHoudiniPublicAPIAssetWrapper::SetRampParameterNumPoints_Implementation(FName In
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -1683,7 +1683,7 @@ UHoudiniPublicAPIAssetWrapper::SetRampParameterNumPoints_Implementation(FName In
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -1897,7 +1897,7 @@ UHoudiniPublicAPIAssetWrapper::GetRampParameterNumPoints_Implementation(FName In
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -1907,7 +1907,7 @@ UHoudiniPublicAPIAssetWrapper::GetRampParameterNumPoints_Implementation(FName In
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -2009,7 +2009,7 @@ UHoudiniPublicAPIAssetWrapper::SetFloatRampParameterPoints_Implementation(
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -2157,7 +2157,7 @@ UHoudiniPublicAPIAssetWrapper::GetFloatRampParameterPoints_Implementation(
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -2279,7 +2279,7 @@ UHoudiniPublicAPIAssetWrapper::SetColorRampParameterPoints_Implementation(
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -2427,7 +2427,7 @@ UHoudiniPublicAPIAssetWrapper::GetColorRampParameterPoints_Implementation(
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -2512,7 +2512,7 @@ UHoudiniPublicAPIAssetWrapper::TriggerButtonParameter_Implementation(FName InBut
 		if (!IsValid(ButtonParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 
@@ -3131,7 +3131,7 @@ UHoudiniPublicAPIAssetWrapper::BakeOutputObjectAt_Implementation(
 		case EHoudiniOutputType::Invalid:
 		default:
 			SetErrorMessage(FString::Printf(
-				TEXT("BakeOutputObjectAt: unsupported output type (%d) for baking to CB."), OutputType));
+				TEXT("BakeOutputObjectAt: unsupported output type (%d) for baking to CB."), static_cast<int32>(OutputType)));
 			return false;
 	}
 
@@ -3770,7 +3770,7 @@ UHoudiniPublicAPIAssetWrapper::FindRampPointData(
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *InParam->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *InParam->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -3780,7 +3780,7 @@ UHoudiniPublicAPIAssetWrapper::FindRampPointData(
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *InParam->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *InParam->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -4037,7 +4037,7 @@ bool UHoudiniPublicAPIAssetWrapper::SetRampParameterPointValue(
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -4047,7 +4047,7 @@ bool UHoudiniPublicAPIAssetWrapper::SetRampParameterPointValue(
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -4254,7 +4254,7 @@ UHoudiniPublicAPIAssetWrapper::GetRampParameterPointValue(
 		if (!IsValid(FloatRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -4264,7 +4264,7 @@ UHoudiniPublicAPIAssetWrapper::GetRampParameterPointValue(
 		if (!IsValid(ColorRampParam))
 		{
 			SetErrorMessage(FString::Printf(
-				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), ParamType));
+				TEXT("Unexpected parameter class (%s) vs type (%d)"), *Param->GetClass()->GetName(), static_cast<int32>(ParamType)));
 			return false;
 		}
 	}
@@ -4472,7 +4472,7 @@ UHoudiniPublicAPIAssetWrapper::CreateAndPopulateAPIInput(const UHoudiniInput* In
 		// Deprecated input types
 		case EHoudiniInputType::Invalid:
 		default:
-			SetErrorMessage(FString::Printf(TEXT("GetInputAtIndex: Invalid input type %d"), InputType));
+			SetErrorMessage(FString::Printf(TEXT("GetInputAtIndex: Invalid input type %d"), static_cast<int32>(InputType)));
 			return false;
 	}
